[tests] Fixed coverage
